requirements.txt
streamlit==1.32.0
numpy==1.26.4
Pillow==10.2.0
opencv-python-headless==4.9.0.80